"Added function to set school_id in User model"
This commit adds a new function called make_school_id to the User model. The function checks if the school_id is set in the current session, and if so, assigns the value to the school_id property of the User object. This allows for easy retrieval and manipulation of the school_id within the User model.