"Remove unwanted columns before insert"
This commit modifies the code to remove unwanted columns before inserting data into the database. The allowedColumns property in the User model is used to specify the list of allowed columns. In this commit, a loop is added to insert method to iterate through the data array and remove any keys that are not present in the allowedColumns list. This ensures that only the desired columns are included in the data being inserted.